+++ sdk/spring/scripts/Update-DevOps-Release-WorkItem/Update-DevOps-Release-WorkItems-For-KeyVaultJCA.ps1
@@ -1,7 +1,7 @@
 #Requires -Version 6.0
 
-$releaseDate = "05/20/2021"
+$releaseDate = "06/23/2021"
 $serviceDirectory = "keyvault"
-$jcaReleaseVersion = "1.0.0-beta.7"
+$jcaReleaseVersion = "1.0.0"
 
 . ${PSScriptRoot}\Update-DevOps-Release-WorkItem-Util.ps1 $releaseDate $serviceDirectory $jcaReleaseVersion azure-security-keyvault-jca